pyproject.toml
[tool.poetry]
name = "blank-blog"
version = "0.1.0"
description = ""
authors = ["haruBlank <harublank00@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.1.1"
django-markdownx = "^4.0.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
